Modified scripts in the package.json file to consider the proper environment variables which different in development, testing and production stage.
Problem: This does not work on Windows. For that we additionally install the cross-env package with:
npm install --save-dev cross-env

and

npm install cross-env

We can then achieve cross-platform compatibility by using the cross-env library in our npm scripts defined in package.json:

Next we changed the config.js file (which is taking over the work of the node-config package that is used for larger applications) to pass different env variables, depending on testing mode or not.
For this, we also have to add a TEST_MONGODB_URI variable in the .env file. (For the time being its the same as the other one)