Finished 4.17: Blog List Expansion, step 5:
- In blog.js:
Extended blogSchema to contain user information for each blog

- In blogs.js:
Refactored POST method for blogs to use async/await syntax instead of Promises.
Also added functionality to add a dummyUserID to a new blog. (This will be replaced with adding the REAL User ID later on)

- Refactored the blogsRouter.get method to use async/await syntax instead of Promises. Then added the populate functionality of mongoose to replace the user: userID field of a blog with the whole user information.

Added functionality:
When a blog is created, the user in the DB is updated to contain a reference to the blog that he created:
- Changed user schema to contain a blogs array field (multiple blogs can be made by one user)
- In blogs.js: The User who created the blog is updated, by adding a reference to the blog to the user.